requirements.txt
Flask>=3.0
azure-data-tables>=12.6
azure-storage-blob>=12.19
pandas>=2.3
openpyxl>=3.1
